+++ trixnity-clientserverapi/trixnity-clientserverapi-model/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/model/rooms/GetPublicRooms.kt
@@ -3,6 +3,8 @@ package net.folivo.trixnity.clientserverapi.model.rooms
 import io.ktor.resources.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import net.folivo.trixnity.core.Auth
+import net.folivo.trixnity.core.AuthRequired
 import net.folivo.trixnity.core.HttpMethod
 import net.folivo.trixnity.core.HttpMethodType.GET
 import net.folivo.trixnity.core.MatrixEndpoint
@@ -13,6 +15,7 @@ import net.folivo.trixnity.core.MatrixEndpoint
 @Serializable
 @Resource("/_matrix/client/v3/publicRooms")
 @HttpMethod(GET)
+@Auth(AuthRequired.NO)
 data class GetPublicRooms(
     @SerialName("limit") val limit: Long? = null,
     @SerialName("server") val server: String? = null,