+++ trixnity-client/src/commonTest/kotlin/net/folivo/trixnity/client/store/cache/ObservableCacheTest.kt
@@ -5,7 +5,6 @@ import io.kotest.core.spec.style.ShouldSpec
 import io.kotest.matchers.collections.shouldContainAll
 import io.kotest.matchers.comparables.shouldBeLessThan
 import io.kotest.matchers.shouldBe
-import korlibs.io.async.launch
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.*
 import kotlinx.datetime.Clock