+++ trixnity-clientserverapi/trixnity-clientserverapi-model/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/model/authentication/GetEmailRequestTokenForRegistration.kt
@@ -3,10 +3,11 @@ package net.folivo.trixnity.clientserverapi.model.authentication
 import io.ktor.resources.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import net.folivo.trixnity.core.Auth
+import net.folivo.trixnity.core.AuthRequired
 import net.folivo.trixnity.core.HttpMethod
 import net.folivo.trixnity.core.HttpMethodType.POST
 import net.folivo.trixnity.core.MatrixEndpoint
-import net.folivo.trixnity.core.WithoutAuth
 
 /**
  * @see <a href="https://spec.matrix.org/v1.10/client-server-api/#post_matrixclientv3registeremailrequesttoken">matrix spec</a>
@@ -14,7 +15,7 @@ import net.folivo.trixnity.core.WithoutAuth
 @Serializable
 @Resource("/_matrix/client/v3/register/email/requestToken")
 @HttpMethod(POST)
-@WithoutAuth
+@Auth(AuthRequired.NO)
 object GetEmailRequestTokenForRegistration :
     MatrixEndpoint<GetEmailRequestTokenForRegistration.Request, GetEmailRequestTokenForRegistration.Response> {
     @Serializable