+++ trixnity-clientserverapi/trixnity-clientserverapi-model/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/model/discovery/GetWellKnown.kt
@@ -3,11 +3,8 @@ package net.folivo.trixnity.clientserverapi.model.discovery
 import io.ktor.resources.*
 import kotlinx.serialization.Serializable
 import net.folivo.trixnity.clientserverapi.model.authentication.DiscoveryInformation
-import net.folivo.trixnity.core.ForceJson
-import net.folivo.trixnity.core.HttpMethod
+import net.folivo.trixnity.core.*
 import net.folivo.trixnity.core.HttpMethodType.GET
-import net.folivo.trixnity.core.MatrixEndpoint
-import net.folivo.trixnity.core.WithoutAuth
 
 /**
  * @see <a href="https://spec.matrix.org/v1.10/client-server-api/#getwell-knownmatrixclient">matrix spec</a>
@@ -15,6 +12,6 @@ import net.folivo.trixnity.core.WithoutAuth
 @Serializable
 @Resource("/.well-known/matrix/client")
 @HttpMethod(GET)
-@WithoutAuth
+@Auth(AuthRequired.NO)
 @ForceJson
 object GetWellKnown : MatrixEndpoint<Unit, DiscoveryInformation>