+++ trixnity-serverserverapi/trixnity-serverserverapi-model/src/commonMain/kotlin/net/folivo/trixnity/serverserverapi/model/discovery/GetServerKeys.kt
@@ -2,10 +2,11 @@ package net.folivo.trixnity.serverserverapi.model.discovery
 
 import io.ktor.resources.*
 import kotlinx.serialization.Serializable
+import net.folivo.trixnity.core.Auth
+import net.folivo.trixnity.core.AuthRequired
 import net.folivo.trixnity.core.HttpMethod
 import net.folivo.trixnity.core.HttpMethodType.GET
 import net.folivo.trixnity.core.MatrixEndpoint
-import net.folivo.trixnity.core.WithoutAuth
 import net.folivo.trixnity.core.model.keys.Signed
 
 /**
@@ -14,5 +15,5 @@ import net.folivo.trixnity.core.model.keys.Signed
 @Serializable
 @Resource("/_matrix/key/v2/server")
 @HttpMethod(GET)
-@WithoutAuth
+@Auth(AuthRequired.NO)
 object GetServerKeys : MatrixEndpoint<Unit, Signed<ServerKeys, String>>