+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fix ConcurrentModificationException on cache removal.
+- Send access token to endpoints when server wants to
 
 ### Security